Added employee service

+++ src/main/kotlin/com/github/npospolita/worktimesheets/domain/WorkReport.kt
@@ -9,6 +9,6 @@ import javax.persistence.Id
 
 @Entity
 data class WorkReport(@Id val id: UUID,
-                      val employeeId: Int,
+                      val employeeId: Long,
                       @ElementCollection val days: List<LocalDate> = ArrayList(),
                       val amount: BigInteger)
